requirements.txt
dbt-snowflake
snowflake-connector-python
pandas
numpy
prophet
matplotlib
streamlit
python-dotenv
faker
mistralai
tabulate
